separated the geoid correction from MobileAdaptor.java and put it into separate class Geoid.java instead.
